(refactor/feature) Finish screenshots/fund report business logic Add Exceptions and handle them; Add types; Create constants for paths;

+++ classes/class.Screenshot.php
@@ -10,7 +10,6 @@ class Screenshot {
   private $recipients_emails;
   private $pages_to_screenshot;
   private $attachments;
-  private $screenshot_path;
   private $screenshots_links;
 
   public function __construct() {
@@ -21,75 +20,84 @@ class Screenshot {
     $this->API_KEY = DB_Options::get('apiflash_api_key');
     $this->recipients_emails = DB_Options::get('screenshots_recipients_emails');
     $this->pages_to_screenshot = DB_Options::get('pages_to_screenshot');
-    $this->screenshot_path = WP_PLUGIN_DIR . '/mst_4f_as/screenshots/';
     $this->screenshots_links = [];
   }
 
-  /**
-   * Used as an entry for the cron task.
-   * */
-  public function take_and_send_all() {
+  public function take_and_send_all(): bool {
     try {
       $this->take_and_save_all(false);
       $this->send_screenshots();
       $this->remove_screenshots();
+
+      return true;
+    } catch (Exception $e) {
+      Logger::log('error', [ 'error' => $e ]);
+      return false;
+    }
+  }
+
+  private function take_and_save_all(bool $is_preload) {
+    try {
+      $pages = explode("\n", str_replace("\r", '', $this->pages_to_screenshot));
+
+      foreach ($pages as $page) {
+        $this->take($page, $is_preload);
+      }
+
+      if ($is_preload) exit;
+
+      $this->download_screenshots();
+
+      if (empty(MST_4F_AS_SCREENSHOTS_PATH)) {
+        throw new Exception('MST_4F_AS_SCREENSHOTS_PATH constant is empty.');
+      }
+
+      $screenshots = array_diff(scandir(MST_4F_AS_SCREENSHOTS_PATH), [ '.', '..' ]);
+
+      if (!sizeof($screenshots)) {
+        Logger::log('warning', [ 'error' => 'There are no screenshots to send.' ]);
+        exit;
+      }
+
+      $this->attachments = array_map(function(string $screenshot): string {
+        return sprintf('%s/%s', MST_4F_AS_SCREENSHOTS_PATH, $screenshot);
+      }, $screenshots);
     } catch (Exception $e) {
       Logger::log('error', [ 'error' => $e ]);
     }
   }
 
-  private function take_and_save_all(bool $is_preload = false) {
-    $pages = explode("\n", str_replace("\r", '', $this->pages_to_screenshot));
+  private function take(string $page_url, bool $fresh = false): array {
+    try {
+      if (empty($this->API_KEY)) {
+        throw new Exception('ApiFlash key must be provided.');
+      }
 
-    foreach ($pages as $page) {
-      $this->take($page, $is_preload);
-    }
+      $request_url = add_query_arg([
+        'access_key' => $this->API_KEY,
+        'url' => urlencode($page_url),
+        'format' => 'jpeg',
+        'full_page' => true,
+        'fresh' => $fresh,
+        'ttl' => 2 * 3600,
+        'delay' => 2,
+        'quality' => 60,
+        'response_type' => 'json'
+      ], 'https://api.apiflash.com/v1/urltoimage');
 
-    if ($is_preload) {
+      $json = json_decode(file_get_contents($request_url));
+
+      if (empty($json)) {
+        throw new Exception('Maximum api call count reached.');
+      }
+
+      $this->screenshots_links[] = $json->url;
+
+      return $this->screenshots_links;
+    } catch (Exception $e) {
+      Logger::log('error', [ 'error' => $e ]);
       exit;
     }
-
-    $this->download_screenshots();
-
-    $screenshots = array_diff(scandir($this->screenshot_path), [ '.', '..' ]);
-
-    if (!sizeof($screenshots)) {
-      return;
-    }
-
-    $this->attachments = array_map(function (string $screenshot) {
-      return $this->screenshot_path . $screenshot;
-    }, $screenshots);
-  }
-
-  private function take(string $page_url, bool $fresh = false): array {
-    if (empty($this->API_KEY)) {
-      Logger::log('error', [ 'error' => 'ERROR: ApiFlash key must be provided' ]);
-      wp_die('ERROR: ApiFlash key must be provided ' . print_r(debug_backtrace(), 1));
-    }
-
-    $request_url = add_query_arg([
-      'access_key' => $this->API_KEY,
-      'url' => urlencode($page_url),
-      'format' => 'jpeg',
-      'full_page' => true,
-      'fresh' => $fresh,
-      'ttl' => 2 * 3600,
-      'delay' => 2,
-      'quality' => 60,
-      'response_type' => 'json'
-    ], 'https://api.apiflash.com/v1/urltoimage');
-
-    $json = json_decode(file_get_contents($request_url));
-
-    if (empty($json)) {
-      Logger::log('error', [ 'error' => 'Maximum api call count reached' ]);
-      wp_die();
-    }
-
-    $this->screenshots_links[] = $json->url;
-
-    return $this->screenshots_links;
   }
 
   private function download_screenshots() {
@@ -99,43 +107,64 @@ class Screenshot {
       $image = file_get_contents($screenshots_link);
 
       file_put_contents(
-        $this->screenshot_path . sprintf('%s--%s.jpeg', $current_datetime, $random_id),
+        MST_4F_AS_SCREENSHOTS_PATH . sprintf('/%s--%s.jpeg', $current_datetime, $random_id),
         $image
       );
     }
   }
 
   private function send_screenshots() {
-    $recipients = explode("\n", str_replace("\r", '', $this->recipients_emails));
-    $date = date('Y-m-d');
+    try {
+      $recipients = explode("\n", str_replace("\r", '', $this->recipients_emails));
+      $date = date('Y-m-d');
 
-    if (!sizeof($this->attachments) || !sizeof($recipients)) return;
+      if (empty($this->attachments)) {
+        throw new Exception('Message is empty.');
+      }
 
-    foreach ($recipients as $recipient) {
-      wp_mail(
-        sanitize_email($recipient),
-        sprintf('[%s] Report of the 4F funds', $date),
-        'Hi! A bunch of new screenshots already here.',
-        '',
-        $this->attachments
-      );
+      if (!sizeof($recipients)) {
+        throw new Exception('There are no recipients for the screenshots.');
+      }
+
+      foreach ($recipients as $recipient) {
+        wp_mail(
+          sanitize_email($recipient),
+          sprintf('[%s] Report of the 4F funds', $date),
+          'Hi! A bunch of new screenshots already here.',
+          '',
+          $this->attachments
+        );
+      }
+    } catch (Exception $e) {
+      Logger::log('error', [ 'error' => $e ]);
     }
   }
 
   private function remove_screenshots() {
-    $files = glob($this->screenshot_path . '*');
+    try {
+      if (empty(MST_4F_AS_SCREENSHOTS_PATH)) {
+        throw new Exception('MST_4F_AS_FUNDS_ARCHIVE_PATH constant is empty.');
+      }
 
-    foreach ($files as $file) {
-      if (is_file($file))
-        unlink($file);
+      $files = glob(MST_4F_AS_SCREENSHOTS_PATH . '/*');
+
+      foreach ($files as $file) {
+        if (is_file($file)) {
+          unlink($file);
+        }
+      }
+    } catch (Exception $e) {
+      Logger::log('error', [ 'error' => $e ]);
     }
   }
 
-  public function preload_screenshots() {
+  public function preload_screenshots(): bool {
     try {
       $this->take_and_save_all(true);
+      return true;
     } catch (Exception $e) {
       Logger::log('error', [ 'error' => $e ]);
+      return false;
     }
   }
 }
